添加脚本管理器

+++ spring-boot3-brick-bootkit-scripts/src/main/java/com/zqzqq/bootkits/scripts/monitor/impl/DefaultScriptPerformanceMonitor.java
@@ -206,7 +206,7 @@ public class DefaultScriptPerformanceMonitor implements ScriptPerformanceMonitor
             .mapToDouble(r -> r.getMetrics().getCpuUsagePercent())
             .average()
             .orElse(0.0);
-        long avgMemoryUsage = filteredRecords.stream()
+        long avgMemoryUsage = (long) filteredRecords.stream()
             .mapToLong(r -> r.getMetrics().getMemoryUsedBytes())
             .average()
             .orElse(0.0);
